+++ myschedule/settings/base.py
@@ -10,8 +10,7 @@ app_root = d(d(__file__))
 APP_STATIC_MEDIA = j(app_root, 'static')
 
 LOGIN_URL = '/myschedule/login/'
-CAS_REDIRECT_URL = '/myschedule/'
-CAS_IGNORE_REFERER = False
+CAS_REDIRECT_URL = '/myschedule/schedule_login/'
 
 APP_NAME = 'myschedule'
 BASE_URL = APP_NAME + '/'